Create _3_password_generator.py
3.	Write a password generator in Python. Be creative with how you generate passwords - strong passwords have a mix of lowercase letters, uppercase letters, numbers, and symbols. The passwords should be random, generating a new password every time the user asks for a new password. Ask the user how strong they want their password to be.
a.	The starting character should be an capital
b.	The second character should be an symbol
c.	The third character should be a prime number
d.	The last number should be a composite number
e.	Password length must be size of 8<=15 characters
f.	If all the characters fail the above condition, then print " not a valid password"