Create _6b_airline_bags.py
Anitha  has 3 bags that she wants to take on a flight. They weigh A, B and C kgs respectively. She wants to
check in exactly two of these bags and carry the remaining one bag with her.The airline restrictions says that the
total sum of the weights of the bags that are checked-in cannot exceed D kgs and the weight of the bag which is
carried cannot exceed E kgs. Find if Anitha can take all the three bags on the flight.